pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-orchestration-platform"
version = "0.1.0"
description = "Unified multi-agent orchestration platform"
requires-python = ">=3.10"
dependencies = ["click>=8.0.0", "pydantic>=2.0.0", "rich>=13.0.0", "pyyaml>=6.0.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-cov>=4.0.0"]

[project.scripts]
aop = "aop.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/aop"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]